feat: npm run import でscripts配下のImportZipsを実行できるようにする
CLIエントリポイントをscripts/ImportZips.jsへ集約し、npm run import -- <dir> で
対象ディレクトリを Query に渡して ImportZips.execute を呼び出せるようにした。
既存の application 配下参照との互換性を維持するため、旧パスは再エクスポートへ変更した。
