+++ package.json
@@ -17,7 +17,8 @@
     "watch-api": "node ./node_modules/aglio/bin/aglio.js -i ./doc/5_api/openapi/openapi.yaml -s",
     "dev:entry": "npm run dev",
     "seed:user": "node ./scripts/seed-fixed-user.js",
-    "start:seeded": "npm run seed:user && npm run start"
+    "start:seeded": "npm run seed:user && npm run start",
+    "import": "node ./scripts/ImportZips.js"
   },
   "author": "",
   "license": "MIT",
